Tweak autocraft language
