+++ datagen/src/main/resources/meta/tinactory/language/en_us.json
@@ -156,8 +156,8 @@
     "tinactory.tooltip.powerBlock": "Capacity: %s EU",
     "tinactory.tooltip.batteryBox": "Battery Slots: %s",
     "tinactory.tooltip.meDrive": "Storage Slots: %s",
-    "tinactory.tooltip.electricChest": "Capacity: %s x %s",
-    "tinactory.tooltip.electricTank": "Capacity: %s mB x %s",
+    "tinactory.tooltip.electricChest": "Capacity: %s ×%s",
+    "tinactory.tooltip.electricTank": "Capacity: %s mB ×%s",
     "tinactory.tooltip.logisticWorker.1": "Max Tasks: %s",
     "tinactory.tooltip.logisticWorker.2": "Task Interval: %s s",
     "tinactory.tooltip.logisticWorker.3": "Transmission Capacity: %s (Items) / %s mB (Fluid)",
@@ -217,8 +217,8 @@
     "tinactory.gui.autocraft.cpu.state.running": "Running",
     "tinactory.gui.autocraft.cpu.state.blocked": "Blocked",
     "tinactory.gui.autocraft.cpu.state.failed": "Failed",
-    "tinactory.gui.autocraft.cpu.target": "%s x %s",
-    "tinactory.gui.autocraft.cpu.steps": "%s / %s",
+    "tinactory.gui.autocraft.cpu.target": "%s ×%s",
+    "tinactory.gui.autocraft.cpu.steps": "Steps: %s / %s",
     "tinactory.gui.autocraft.cpu.error.offline": "CPU service is offline",
     "tinactory.gui.autocraft.cpu.error.inputUnavailable": "Input resources are unavailable",
     "tinactory.gui.autocraft.cpu.error.machineUnavailable": "Machine requirement is unavailable",
@@ -227,7 +227,7 @@
     "tinactory.gui.pattern.input": "Input: %s",
     "tinactory.gui.pattern.output": "Output: %s",
     "tinactory.gui.pattern.empty": "<Empty>",
-    "tinactory.gui.pattern.ingredient": "%s x %s",
+    "tinactory.gui.pattern.ingredient": "%s ×%s",
     "tinactory.gui.pattern.ingredients": "%s %s",
     "tinactory.gui.pattern.ellipsis": "...",
     "tinactory.gui.pattern.editor.input": "Input",
