+++ datagen/src/main/resources/meta/tinactory/language/zh_cn.json
@@ -550,8 +550,8 @@
     "tinactory.tooltip.powerBlock": "容量：%s EU",
     "tinactory.tooltip.batteryBox": "电池槽：%s",
     "tinactory.tooltip.meDrive": "存储槽：%s",
-    "tinactory.tooltip.electricChest": "容量：%s x %s",
-    "tinactory.tooltip.electricTank": "容量：%s mB x %s",
+    "tinactory.tooltip.electricChest": "容量：%s ×%s",
+    "tinactory.tooltip.electricTank": "容量：%s mB ×%s",
     "tinactory.tooltip.logisticWorker.1": "最大任务数量：%s",
     "tinactory.tooltip.logisticWorker.2": "任务周期：%s s",
     "tinactory.tooltip.logisticWorker.3": "传输容量：%s (Items) / %s mB (Fluid)",
@@ -611,8 +611,8 @@
     "tinactory.gui.autocraft.cpu.state.running": "运行中",
     "tinactory.gui.autocraft.cpu.state.blocked": "阻塞",
     "tinactory.gui.autocraft.cpu.state.failed": "失败",
-    "tinactory.gui.autocraft.cpu.target": "%s x%s",
-    "tinactory.gui.autocraft.cpu.steps": "%s / %s",
+    "tinactory.gui.autocraft.cpu.target": "%s ×%s",
+    "tinactory.gui.autocraft.cpu.steps": "步骤：%s / %s",
     "tinactory.gui.autocraft.cpu.error.offline": "CPU 服务离线",
     "tinactory.gui.autocraft.cpu.error.inputUnavailable": "输入资源不可用",
     "tinactory.gui.autocraft.cpu.error.machineUnavailable": "机器需求不可用",
@@ -621,7 +621,7 @@
     "tinactory.gui.pattern.input": "输入: %s",
     "tinactory.gui.pattern.output": "输出: %s",
     "tinactory.gui.pattern.empty": "<空>",
-    "tinactory.gui.pattern.ingredient": "%s x %s",
+    "tinactory.gui.pattern.ingredient": "%s ×%s",
     "tinactory.gui.pattern.ingredients": "%s %s",
     "tinactory.gui.pattern.ellipsis": "...",
     "tinactory.gui.pattern.editor.input": "输入",
